feat: add status-badges component in GN2CommonModule
