+++ frontend/src/app/GN2CommonModule/others/status-badges/status-badges.component.html
@@ -1,0 +1,8 @@
+<div class="StatusBadges">
+  <gn-badge
+    *ngFor="let status of status"
+    [tooltip]="status.tooltip"
+    [text]="status.badge"
+    [symbology]="status.symbology"
+  />
+</div>
