+++ frontend/src/app/GN2CommonModule/others/status-badges/status-badges.component.scss
@@ -1,0 +1,7 @@
+.StatusBadges {
+  display: flex;
+  flex-flow: row wrap;
+  width: 100%;
+  gap: 0.5rem;
+  line-height: inherit;
+}
